Move containers into own file
